Fix b200 failures: welford cute bool-sum NaN + reduction-heuristic test cache pollution
  examples/welford.py: cast the valid-column mask to int32 before .sum().
  The CuTe backend derives a reduction's accumulator dtype from its operand,
  so summing a bool mask used a saturating boolean accumulator (Tn collapsed
  to 1 -> wrong mean -> negative variance -> NaN). The int32 cast forces an
  integer accumulator on every backend and preserves the non-divisor-N fix.

  test/test_autotuner_heuristics.py: reset() the kl_div_forward / rms_norm_fwd
  singletons before binding under the HOPPER patch. bind() caches BoundKernels
  by arg signature, not hardware, so a prior non-patched bind (or a --reruns
  retry) on the same xdist worker returned a stale BoundKernel with
  autotuner_heuristics computed for the real (non-sm90) device, silently
  bypassing the patch. Matches the existing reset() idiom in test_cache.py.

+++ examples/welford.py
@@ -56,7 +56,12 @@ def welford(
             # over-counts the padding on the last tile when block_size does not
             # divide n -> wrong mean/M2/count at non-divisor N). `tile_n.index`
             # is the canonical Helion masked-index idiom (tile_interface.index).
-            Tn = (tile_n.index < n).sum()
+            # Cast the bool mask to int32 BEFORE summing: the CuTe backend derives
+            # the reduction accumulator dtype from the operand, and a bool operand
+            # yields a saturating boolean accumulator (Tn collapses to 1, not the
+            # true count -> wrong mean -> negative variance -> NaN). The explicit
+            # int32 cast forces an integer accumulator on every backend.
+            Tn = (tile_n.index < n).to(torch.int32).sum()
             sum_x = torch.sum(chunk, dim=-1)
             sum_x2 = torch.sum(chunk * chunk, dim=-1)
             mean_c = sum_x / Tn
